lib/ruby combine_ranges.rb
Takes an Array of integer ranges represented as a two element array where the first element is the range min (begin) and the 2nd/last is the end of the range.  The method combines all overlapping entries.  It also combines those entries with the max value of one range is one less than the min balue of another range.